src/gprofiler-dev/postgres_requirements.txt


psycopg2-binary==2.9.9
pydantic[email]==1.8.2
